requirements.txt
Flask
flask-cors
python-pptx
python-docx
PyMuPDF
openai==0.27.0
gunicorn
